requirements_apple_silicon.txt
# TensorFlow for Apple Silicon (M1/M2/M3)
tensorflow-macos==2.12.0
tensorflow-metal==1.0.0

# Core dependencies
numpy>=1.23.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pillow>=9.0.0
h5py>=3.7.0

# Computer Vision
opencv-python>=4.5.0

# Additional ML utilities
pandas>=1.5.0

# Configuration management
PyYAML>=6.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0
pytest-xdist>=3.0.0  # For parallel test execution
psutil>=5.9.0  # For performance monitoring in tests

# Note: Use Python 3.8-3.11 for best compatibility with tensorflow-macos